+++ assets/training/model_evaluation/src/task_factory/base.py
@@ -103,6 +103,7 @@ class BasePredictor(ABC):
 
     def handle_device_failure(self, X_test, **kwargs):
         """Handle device failure."""
+        predict_fn = kwargs.get('predict_fn', self.model.predict)
         if self.device == constants.DEVICE.AUTO and torch.cuda.is_available():
             try:
                 cuda_current_device = torch.cuda.current_device()
@@ -112,26 +113,32 @@ class BasePredictor(ABC):
                         f"Current Device: {self.current_device} does not match expected device {cuda_current_device}")
                     self.model = load_model(self.model_uri, cuda_current_device, self.task_type)
                     self.current_device = cuda_current_device
+                    predict_fn = kwargs.get('predict_fn', self.model.predict)
                 kwargs["device"] = self.current_device
-                return self.model.predict(X_test, **kwargs)
-            except TypeError:
-                return self.model.predict(X_test)
+                return predict_fn(X_test, **kwargs)
+            except TypeError as e:
+                logger.warning("Failed on GPU with error: " + repr(e))
+                logger.info("Trying on GPU without kwargs.")
+                return predict_fn(X_test)
             except Exception as e:
-                logger.info("Failed on GPU with error: " + repr(e))
+                logger.warning("Failed on GPU with error: " + repr(e))
         if self.device != -1:
             logger.warning("Running predictions on CPU.")
             self.model = load_model(self.model_uri, -1, self.task_type)
+            predict_fn = kwargs.get('predict_fn', self.model.predict)
             try:
                 logger.info("Loading model and prediction with cuda current device. Trying CPU.")
                 if 'device' in kwargs:
                     kwargs.pop('device')
                 if 'device_map' in kwargs:
                     kwargs.pop('device_map')
-                return self.model.predict(X_test, **kwargs)
-            except TypeError:
-                return self.model.predict(X_test)
+                return predict_fn(X_test, **kwargs)
+            except TypeError as e:
+                logger.warning("Failed on CPU with error: " + repr(e))
+                logger.info("Trying on CPU without kwargs.")
+                return predict_fn(X_test)
             except Exception as e:
-                logger.info("Failed on CPU with error: " + repr(e))
+                logger.warning("Failed on CPU with error: " + repr(e))
                 raise e