+++ assets/training/model_management/src/azureml/model/mgmt/processors/pyfunc/vision/config.py
@@ -32,6 +32,7 @@ class MMDetLiterals:
     WEIGHTS_PATH = "weights_path"
     AUGMENTATIONS_PATH = "augmentations_path"
     METAFILE_PATH = "model_metadata"
+    CLASSES = "classes"
 
 
 class MLflowSchemaLiterals:
@@ -68,7 +69,7 @@ class ODLiterals:
     BOTTOM_Y = "bottomY"
 
 
-class ISLiterals(ODLiterals):
+class ISLiterals:
     """IS constants."""
 
     POLYGON = "polygon"