+++ scripts/azureml-assets/azureml/assets/validate_assets.py
@@ -4,12 +4,15 @@
 """Validate assets."""
 
 import argparse
+import json
 import re
 import sys
 from collections import defaultdict
 from pathlib import Path
 from typing import List
 
+from azure.ai.ml import load_model
+from azure.ai.ml.operations._run_history_constants import JobStatus
 import azureml.assets as assets
 import azureml.assets.util as util
 from azureml.assets import PublishLocation, PublishVisibility
@@ -30,6 +33,10 @@ ASSET_CONFIG_URL = "https://github.com/Azure/azureml-assets/wiki/Assets#assetyam
 
 # Model naming convention
 MODEL_NAME_PATTERN = re.compile(r"^[A-Za-z][A-Za-z0-9_.-]{0,254}$")
+
+# model validations
+MODEL_VALIDATION_RESULTS_FOLDER = "validation_results"
+MODEL_VALIDATION_JOB_DETAILS = "job_details.json"
 
 # Environment naming convention
 ENVIRONMENT_NAME_PATTERN = re.compile(r"^[a-z][a-z0-9.-]{0,254}$")
@@ -313,8 +320,114 @@ def validate_categories(asset_config: assets.AssetConfig) -> int:
     return 0
 
 
+def validate_model_assets(latest_asset_config: assets.AssetConfig, validated_asset_config: assets.AssetConfig) -> int:
+    """Check if current model asset and validated one matches and has a successful run."""
+    # latest_asset_config is expected to be non null
+    if not validated_asset_config:
+        logger.log_error(f"Validated asset config is None for {latest_asset_config.name}")
+        return 1
+
+    logger.print(f"Comparing validated and latest model asset files for {latest_asset_config.name}")
+
+    try:
+        # check if spec has changes
+        latest_model = load_model(latest_asset_config.spec_with_path)
+        validated_model = load_model(validated_asset_config.spec_with_path)
+
+        if latest_model.version != validated_model.version:
+            logger.log_error("version mismatch")
+            logger.log_warning(f"latest_model tags: [[{latest_model.version}]]")
+            logger.log_warning(f"validated_model: [[{validated_model.version}]]")
+            return 1
+
+        if latest_model.tags != validated_model.tags:
+            logger.log_error("tags mismatch")
+            logger.log_warning(f"latest_model tags: [{latest_model.tags}]")
+            logger.log_warning(f"validated_model: [{validated_model.tags}]")
+            return 1
+
+        if latest_model.properties != validated_model.properties:
+            logger.log_error("properties mismatch")
+            logger.log_warning(f"latest_model properties: [{latest_model.properties}]")
+            logger.log_warning(f"validated_model properties: [{validated_model.properties}]")
+            return 1
+
+        if latest_model.description != validated_model.description:
+            logger.log_error("description mismatch")
+            logger.log_warning(f"latest_model description: [{latest_model.description}]")
+            logger.log_warning(f"validated_model description: [{validated_model.description}]")
+            return 1
+
+        latest_model_config: assets.ModelConfig = latest_asset_config.extra_config_as_object()
+        validated_model_config: assets.ModelConfig = validated_asset_config.extra_config_as_object()
+        if not (
+            latest_model_config.path.type == validated_model_config.path.type and
+            latest_model_config.path.uri == validated_model_config.path.uri and
+            latest_model_config.description == validated_model_config.description and
+            latest_model_config.type == validated_model_config.type
+        ):
+            logger.log_error(
+                "Validated model config does not match with latest model. "
+                "Either last validation run for model had failed or its still running."
+            )
+            logger.log_warning(f"latest_model_config: [{latest_model_config._yaml}]")
+            logger.log_warning(f"validated_model_config: [{validated_model_config._yaml}]")
+            if latest_model_config.description != validated_model_config.description:
+                logger.log_warning("Description does not match, for latest and validated asset")
+            return 1
+
+        # check validation results now
+        validation_results_dir = validated_asset_config.file_path / MODEL_VALIDATION_RESULTS_FOLDER
+        validation_job_details_path = validation_results_dir / MODEL_VALIDATION_JOB_DETAILS
+        if not validation_job_details_path.exists():
+            logger.log_error(
+                f"{MODEL_VALIDATION_JOB_DETAILS} missing for model {latest_asset_config.name}. "
+                "Either last validation run for model had failed or its still running."
+            )
+            return 1
+
+        with open(validation_job_details_path) as f:
+            job_details = json.load(f)
+            run_status = job_details.get("status", JobStatus.NOT_STARTED)
+            if run_status != JobStatus.COMPLETED:
+                logger.log_error(
+                    f"run status for model {latest_asset_config.name} is {run_status}. "
+                    "Please ensure that there is a completed model validation job."
+                )
+                return 1
+
+        return 0
+    except Exception as e:
+        logger.log_error(
+            f"Exception when confirming validation results for model {latest_asset_config.name}. Exception {e}"
+        )
+        return 1
+
+
+def get_validated_models_assets_map(model_validation_results_dir: str):
+    """Return model assets map."""
+    try:
+        if not model_validation_results_dir:
+            logger.log_warning(
+                "Unexpected !!! model_validation_results_dir is None. Model assets might fail in validation."
+            )
+            return {}
+
+        validated_model_assets: List[assets.AssetConfig] = util.find_assets(
+            input_dirs=[Path(model_validation_results_dir)],
+            asset_config_filename=assets.DEFAULT_ASSET_FILENAME,
+            types=[assets.config.AssetType.MODEL]
+        )
+
+        return {model_asset.name: model_asset for model_asset in validated_model_assets}
+    except Exception as e:
+        logger.log_error(f"Error in creating validated model map => {e}")
+        return {}
+
+
 def validate_assets(input_dirs: List[Path],
                     asset_config_filename: str,
+                    model_validation_results_dir: str = None,
                     changed_files: List[Path] = None,
                     check_names: bool = False,
                     check_names_skip_pattern: re.Pattern = None,
@@ -326,6 +439,7 @@ def validate_assets(input_dirs: List[Path],
     Args:
         input_dirs (List[Path]): Directories containing assets.
         asset_config_filename (str): Asset config filename to search for.
+        model_validation_results_dir (str, optional): Dir containing model validation results
         changed_files (List[Path], optional): List of changed files, used to filter assets. Defaults to None.
         check_names (bool, optional): Whether to check asset names. Defaults to False.
         check_names_skip_pattern (re.Pattern, optional): Regex pattern to skip name validation. Defaults to None.
@@ -341,6 +455,7 @@ def validate_assets(input_dirs: List[Path],
     """
     # Gather list of just changed assets, for later filtering
     changed_assets = util.find_asset_config_files(input_dirs, asset_config_filename, changed_files) if changed_files else None  # noqa: E501
+    validated_model_map = get_validated_models_assets_map(model_validation_results_dir)
 
     # Find assets under input dirs
     asset_count = 0
@@ -365,6 +480,10 @@ def validate_assets(input_dirs: List[Path],
 
         # Populate dictionary of asset names to asset config paths
         asset_dirs[f"{asset_config.type.value} {asset_config.name}"].append(asset_config_path)
+
+        # validated_model_map would be ampty for non-drop scenario
+        if validated_model_map and asset_config.type == assets.AssetType.MODEL:
+            error_count += validate_model_assets(asset_config, validated_model_map.get(asset_config.name, None))
 
         # Populate dictionary of image names to asset config paths
         environment_config = None
@@ -445,6 +564,8 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument("-i", "--input-dirs", required=True,
                         help="Comma-separated list of directories containing assets")
+    parser.add_argument("-m", "--model-validation-results-dir", required=False,
+                        help="Dir containing model validation results and validated spec.")
     parser.add_argument("-a", "--asset-config-filename", default=assets.DEFAULT_ASSET_FILENAME,
                         help="Asset config file name to search for")
     parser.add_argument("-c", "--changed-files",
@@ -481,6 +602,7 @@ if __name__ == '__main__':
                               check_names_skip_pattern=args.check_names_skip_pattern,
                               check_images=args.check_images,
                               check_categories=args.check_categories,
-                              check_build_context=args.check_build_context)
+                              check_build_context=args.check_build_context,
+                              model_validation_results_dir=args.model_validation_results_dir)
     if not success:
         sys.exit(1)